Make sure to be truly at home
On at least the default FreeBSD install, `/home` is actually a symlink
to `/usr/home`. Because of that, when we open a new terminal we end up
landing at `/usr/home/<username>` instead of `/home/<username>`. Thus
the command prompt doesn't show us `~` when we are home, which is a
useful indicator of where we are.