Set xscreensaver DPMS values a bit higher
We kept blanking the screen too quickly with the previous settings. I had
understood the DPMS settings to be relative to when the screen was
blanked/locked, but it seems that rather, they are measured are in absolute
times from last detected activity.

Now DPMS kicks in one minute before the screen locks, giving us ample time to
quickly wake it up before needing to do a true unlock.