Remove vdirsyncer calls from crontab
I haven't used its output in a long time, and it was constantly erroring
and sending spurious emails to me. Its in the git history if I ever want
to resurrect its use.

Also, only back up the password database on reboot. It hasn't been updated
in a while and running that script hourly is a waste of resources better
spent elsewhere.

Reduce the wait time on reboot cron entries to 30 minutes instead
of a full hour.

Have syncthing log elsewhere and make sure the directory it logs to exists.