Remove aggressive-resize option from tmux
It breaks support in iTerm2.

Remove tmux alias.

Add vscode configuration. Since this lives in the root of the home
directory, I was initially worried that it might conflict with the
global settings for vscode, but it seems that the application buries
the settings sufficiently deep in the home directory on any relevant
systems that it shouldn't be an issue for the repo:

    * https://code.visualstudio.com/docs/getstarted/settings#_settings-file-locations