Improve completion menu experience in vim
Do not auto-select an item in the completion menu anymore and only
insert an item when it is explicitly selected in the menu. This allows
hitting `<CR>` when writing (for example) Markdown without inserting a
random completion. Only if `<Tab>` (or similar) is used to select an
item will `<CR>` cause it to be inserted into the buffer.